+++ src/timessquare/main.py
@@ -15,7 +15,6 @@ from contextlib import asynccontextmanager
 from importlib.metadata import version
 from pathlib import Path
 
-import sentry_sdk
 from fastapi import FastAPI
 from fastapi.openapi.utils import get_openapi
 from safir.database import create_database_engine, is_database_current
@@ -23,7 +22,6 @@ from safir.dependencies.db_session import db_session_dependency
 from safir.fastapi import ClientRequestError, client_request_error_handler
 from safir.logging import configure_logging, configure_uvicorn_logging
 from safir.middleware.x_forwarded import XForwardedMiddleware
-from safir.sentry import before_send_handler
 from safir.slack.webhook import SlackRouteErrorHandler
 from structlog import get_logger
 
@@ -32,15 +30,12 @@ from .dependencies.requestcontext import context_dependency
 from .handlers.external import external_router
 from .handlers.internal import internal_router
 from .handlers.v1 import v1_router
-from .sentry import make_traces_sampler
+from .sentry import init_sentry, make_traces_sampler
 
 __all__ = ["app", "config"]
 
-sentry_sdk.init(
-    dsn=config.sentry_dsn,
-    environment=config.environment_name,
-    before_send=before_send_handler,
-    traces_sampler=make_traces_sampler(config.sentry_traces_sample_rate),
+init_sentry(
+    traces_sampler=make_traces_sampler(config.sentry_traces_sample_rate)
 )